internal/gamepaddb: don't use an Android-specific test file name
The test file name doesn't have to be platform-specific when there is no
platform-specific source file, so put the test in gamepaddb_test.go.
